backend/requirements_save.txt
asgiref==3.9.1
decorator==5.2.1
dj-database-url==3.0.1
dj.choices==0.11.0
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.2
django-typomatic==2.5.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.6.3
whitenoise==6.9.0
gunicorn